pyproject.toml
[tool.poetry]
name = "synthdoc"
version = "0.1.0"
description = "A comprehensive library for generating synthetic documents designed for training and evaluating models in document understanding tasks"
authors = ["Adithya Skolavi <adithya@example.com>"]
readme = "README.md"
packages = [{include = "synthdoc"}]
license = "Apache-2.0"
homepage = "https://github.com/yourusername/synthdoc"
repository = "https://github.com/yourusername/synthdoc"
keywords = ["document", "synthesis", "ocr", "machine-learning", "nlp", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

[tool.poetry.dependencies]
python = "^3.12"
# Core dependencies
pillow = "^10.0.0"
numpy = "^1.24.0"
pydantic = "^2.5.0"
tqdm = "^4.66.0"
requests = "^2.31.0"

# Document processing
reportlab = "^4.0.4"
pypdf = "^3.17.0"
pdfplumber = "^0.10.0"
markdown = "^3.5.1"
beautifulsoup4 = "^4.12.2"

# Image processing
opencv-python = "^4.8.0"
scikit-image = "^0.21.0"

# ML and datasets
datasets = "^2.14.0"
transformers = "^4.35.0"
torch = {version = "^2.1.0", optional = true}
huggingface_hub = "^0.19.0"

# Font handling
fonttools = "^4.43.0"

# CLI
typer = "^0.9.0"
rich = "^13.7.0"

# Environment management
python-dotenv = "^1.0.0"

# Optional LLM integration
openai = {version = "^1.3.0", optional = true}
litellm = {version = "^1.44.0", optional = true}

[tool.poetry.extras]
llm = ["openai", "litellm"]
torch = ["torch"]
full = ["openai", "litellm", "torch"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.9.0"
isort = "^5.12.0"

[tool.poetry.scripts]
synthdoc = "synthdoc.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=synthdoc --cov-report=html --cov-report=term-missing"
